pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibe"
version = "0.1.0"
description = "Vibe coding supervisor without battaries"
readme = "README.md"
authors = [
  {name = "Denis Timofeev", email = "detijazzz@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "pydantic-settings",
    "fastapi",
    "uvicorn",
    "click",
]

[dependency-groups]
dev = [
    "anyio",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.urls]
bugs = "https://github.com/deti/vibe"

[project.scripts]
show-settings = "vibe.cli.show_settings:main"
serve = "vibe.cli.serve:main"

# Configure hatchling to find packages in src/
[tool.hatch.build.targets.wheel]
packages = ["src/vibe"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/pyproject.toml",
]

# Tool configurations
[tool.ruff]
src = ["src", "tests"]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg",
    ".mypy_cache", ".nox", ".pants.d", ".pytype", ".ruff_cache",
    ".svn", ".tox", ".venv", "__pypackages__", "_build",
    "buck-out", "build", "dist", "node_modules", "venv",
]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E4", "E7", "E9", "F", "W", "B", "I", "N", "UP", "C4",
    "ICN", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET",
    "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA",
    "PGH", "PL", "TRY", "NPY", "RUF",
]
ignore = [
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value used in comparison
    "N999", # Hyphen in GitHub repo name
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = false
lines-after-imports = 2
known-first-party = ["vibe"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["RUF001"]
